# mongo/server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: NanoXplore RH — API Backend
  version: 1.0.0
  description: |
    API REST du backend NanoXplore RH (Express + MongoDB).
    Authentification par cookie JWT httpOnly (`token`).
    Base URL de montage : `/api` — le serveur écoute sur `http://localhost:3000` par défaut (PORT env).
servers:
  - url: /api
    description: Express API (chemin relatif au serveur)
security: []  # défaut public — chaque opération précise sa sécurité
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: |
        JWT signé HS256, émis par POST /auth/login.
        Durée : 8h par défaut, 30j si remember=true.
        Cookie httpOnly, secure (production), sameSite=strict.

# =============================================================================
# SHARED PARAMETERS
# =============================================================================
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Numéro de page (base 1)
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 100
      description: Nombre d'éléments par page (max 100)
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: ObjectId MongoDB (24 caractères hexadécimaux)

# =============================================================================
# SCHEMAS
# =============================================================================
  schemas:
    Error:
      type: object
      description: Format d'erreur standard — toutes les réponses 4xx/5xx.
      required: [error]
      properties:
        error:
          type: string
          description: Message d'erreur en français
      example:
        error: "Utilisateur introuvable"
    PaginatedResponse:
      type: object
      description: Enveloppe de réponse pour tous les endpoints de liste.
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items: {}
          description: Tableau d'objets de la page courante
        total:
          type: integer
          description: Nombre total d'éléments (toutes pages confondues)
        page:
          type: integer
          description: Page courante
        limit:
          type: integer
          description: Taille de page demandée
    User:
      type: object
      description: |
        Utilisateur de l'application. passwordHash et ldapDn ne sont jamais retournés.
        Le champ `id` (string) remplace `_id` dans la réponse de GET /auth/me.
      properties:
        _id:
          type: string
          description: ObjectId MongoDB
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, hr, director, manager, employee]
          description: |
            Rôles actifs : admin, hr, manager, employee.
            director : rôle legacy — traité comme manager dans l'architecture cible.
        department:
          type: string
          nullable: true
          enum:
            - Engineering
            - Product
            - Design
            - Data
            - Security
            - Infrastructure
            - Finance
            - Legal
            - HR
            - Sales
            - Marketing
            - "Customer Success"
            - Operations
            - Executive
        position:
          type: string
          nullable: true
          maxLength: 150
        managerId:
          type: string
          nullable: true
          description: ObjectId du manager direct (ref User)
        authSource:
          type: string
          enum: [local, ldap]
          description: "local = compte créé manuellement ; ldap = authentification via annuaire"
        locale:
          type: string
          enum: [fr, en]
        theme:
          type: string
          enum: [dark, light, light-sidebar]
        notificationPrefs:
          type: object
          description: |
            Clés booléennes filtrées par rôle.
            Clés possibles : campaignLaunch, evaluationAssigned, evaluationSubmitted,
            deadlineReminder, managerActionRequired, systemAlerts.
        isActive:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        onboarding:
          type: object
          properties:
            completed:
              type: boolean
            completedAt:
              type: string
              format: date-time
              nullable: true
            steps:
              type: array
              items:
                type: object
                properties:
                  step:
                    type: string
                  done:
                    type: boolean
                  doneAt:
                    type: string
                    format: date-time
                    nullable: true
        offboardingStatus:
          type: string
          enum: [active, offboarding, offboarded]
        offboardingDate:
          type: string
          format: date-time
          nullable: true
        offboardingReason:
          type: string
          nullable: true
        archivedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Form:
      type: object
      description: |
        Formulaire de questions. Une fois qu'une évaluation existe sur ce form,
        frozenAt est renseigné et les questions ne peuvent plus être modifiées.
        formType=upward_feedback force isAnonymous=true.
      properties:
        _id:
          type: string
        campaignId:
          type: string
          nullable: true
          description: ObjectId de la campagne (null = template autonome)
        title:
          type: string
          minLength: 3
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        formType:
          type: string
          enum: [self_evaluation, manager_evaluation, upward_feedback, director_evaluation, peer_review]
        isAnonymous:
          type: boolean
          description: Forcé true si formType=upward_feedback
        questions:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: Identifiant unique de la question dans le form
              type:
                type: string
                enum: [rating, text, yes_no, choice, weather, mobility, n1_import]
              label:
                type: string
                maxLength: 500
              required:
                type: boolean
                default: true
              scale:
                type: integer
                minimum: 2
                maximum: 10
                description: "Uniquement pour type=rating (défaut : 5)"
              options:
                type: array
                items:
                  type: string
                description: "Uniquement pour type=choice (minimum 2 options)"
              phase:
                type: string
                enum: [self, "n-1", objectives, aspirations, all]
                default: all
        frozenAt:
          type: string
          format: date-time
          nullable: true
          description: Renseigné dès la première évaluation — questions gelées
        createdBy:
          description: Peuplé dans les réponses GET
          oneOf:
            - type: string
            - $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Evaluation:
      type: object
      description: |
        Évaluation individuelle (formulaire rempli).
        Cycle de vie : assigned → in_progress → submitted → reviewed
          → signed_evaluatee → signed_manager → signed_hr → validated.
        Statuts terminaux : validated, expired, archived.
        evaluatorId est null dans la réponse si formId.isAnonymous=true.
      properties:
        _id:
          type: string
        campaignId:
          description: Campagne parente (peuplée dans les réponses GET)
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Campaign'
        formId:
          description: Formulaire associé (peuplé dans les réponses GET)
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Form'
        evaluatorId:
          nullable: true
          description: Évaluateur — null si form anonyme (peuplé dans les réponses GET)
          oneOf:
            - type: string
            - $ref: '#/components/schemas/User'
        evaluateeId:
          description: Évalué (peuplé dans les réponses GET)
          oneOf:
            - type: string
            - $ref: '#/components/schemas/User'
        status:
          type: string
          enum: [assigned, in_progress, submitted, reviewed, signed_evaluatee, signed_manager, signed_hr, validated, expired, archived]
        answers:
          type: array
          description: Réponses — bloquées après soumission (LOCKED_STATUSES)
          items:
            type: object
            properties:
              questionId:
                type: string
              value:
                description: "Type Mixed : string, number, boolean, array"
        lastSavedAt:
          type: string
          format: date-time
          nullable: true
        score:
          type: number
          minimum: 0
          maximum: 100
          nullable: true
          description: Score global (0-100), ajouté par le reviewer
        reviewerComment:
          type: string
          maxLength: 5000
        nextObjectives:
          type: string
          maxLength: 5000
        objectiveRatings:
          type: object
          description: "{ [questionId]: 'achieved'|'partial'|'not_achieved' }"
        evaluateeComment:
          type: string
          maxLength: 5000
        disagreementFlag:
          type: boolean
        signedByEvaluateeAt:
          type: string
          format: date-time
          nullable: true
        signedByManagerAt:
          type: string
          format: date-time
          nullable: true
        signedByHrAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: campaign.endDate + 30 jours — le scheduler expire automatiquement
        nearExpiry:
          type: boolean
          description: true si expiresAt < 7 jours (mis par le scheduler)
        auditLog:
          type: array
          description: Journal append-only des actions RH (réaffectations, etc.)
          items:
            type: object
            properties:
              action:
                type: string
              by:
                type: string
              at:
                type: string
                format: date-time
              meta:
                type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /forms:
    get:
      summary: Liste des formulaires
      description: Retourne les formulaires paginés, filtrables par campaignId et/ou formType.
      tags: [forms]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: campaignId
          in: query
          schema:
            type: string
          description: Filtre par campagne (ObjectId)
        - name: formType
          in: query
          schema:
            type: string
            enum: [self_evaluation, manager_evaluation, upward_feedback, director_evaluation, peer_review]
      responses:
        '200':
          description: Liste paginée de formulaires
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Form'
        '400':
          description: campaignId ou formType invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Créer un formulaire
      description: |
        Crée un formulaire lié à une campagne (ou template autonome si campaignId absent).
        formType=upward_feedback force isAnonymous=true.
        Les IDs de questions doivent être uniques dans le formulaire.
      tags: [forms]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, formType]
              properties:
                title:
                  type: string
                  minLength: 3
                  maxLength: 200
                description:
                  type: string
                formType:
                  type: string
                  enum: [self_evaluation, manager_evaluation, upward_feedback, director_evaluation, peer_review]
                campaignId:
                  type: string
                  description: ObjectId (absent = template autonome)
                isAnonymous:
                  type: boolean
                  description: Forcé true si formType=upward_feedback
                questions:
                  type: array
                  items:
                    type: object
      responses:
        '201':
          description: Formulaire créé (peuplé avec createdBy)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form'
        '400':
          description: Validation échouée (title/formType manquants, IDs en double, options choice < 2)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Réservé aux admins et RH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /forms/{id}:
    get:
      summary: Détail d'un formulaire
      description: Retourne un formulaire par son ID.
      tags: [forms]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Formulaire trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form'
        '400':
          description: ID invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Formulaire introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Modifier un formulaire
      description: |
        Modifie un formulaire (admin/hr). title et description toujours modifiables.
        questions : bloquées si frozenAt est défini (retourne 409).
      tags: [forms]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                questions:
                  type: array
                  description: Bloqué si frozenAt est défini
                  items:
                    type: object
      responses:
        '200':
          description: Formulaire modifié (retourne {id} uniquement)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        '400':
          description: Validation Mongoose échouée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Réservé aux admins et RH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Formulaire introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Questions gelées — évaluations existantes sur ce formulaire
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      frozenAt:
                        type: string
                        format: date-time

    delete:
      summary: Supprimer un formulaire
      description: Supprime un formulaire non gelé (admin/hr). Bloqué si frozenAt est défini.
      tags: [forms]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Formulaire supprimé
        '403':
          description: Réservé aux admins et RH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Formulaire introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Formulaire gelé — suppression impossible
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      frozenAt:
                        type: string
                        format: date-time

  # ---------------------------------------------------------------------------
  # Evaluations
  # ---------------------------------------------------------------------------
...
  /evaluations/export:
    get:
      summary: Export CSV des évaluations
      description: Génère un CSV BOM UTF-8 (max 5000 lignes), filtrables par campagne, statut, département (admin/hr).
      tags: [evaluations]
      security:
        - cookieAuth: []
      parameters:
        - name: campaignId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [assigned, in_progress, submitted, reviewed, signed_evaluatee, signed_manager, signed_hr, validated, expired, archived]
        - name: dept
          in: query
          schema:
            type: string
          description: Filtre post-traitement par département (exact match)
      responses:
        '200':
          description: Fichier CSV (Content-Disposition attachment, filename=evaluations.csv)
          content:
            text/csv:
              schema:
                type: string
        '403':
          description: Réservé aux admins et RH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /evaluations/{id}:
    get:
      summary: Détail d'une évaluation
      description: |
        Retourne une évaluation peuplée (form, evaluator, evaluatee, campaign).
        evaluatorId/evaluatorName anonymisés si form.isAnonymous=true.
        Contrôle d'accès par rôle et scope de visibilité.
      tags: [evaluations]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Évaluation trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evaluation'
        '400':
          description: ID invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Accès refusé (hors scope)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Évaluation introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Modifier une évaluation
      description: |
        Sauvegarde les réponses, modifie le statut, ajoute score/commentaires.
        Accès selon le rôle : employee → ses propres évaluations ; manager → ses évaluées.
        Transitions de statut validées par ROLE_TRANSITIONS (voir modèle Evaluation).
        answers verrouillées une fois soumises (LOCKED_STATUSES).
      tags: [evaluations]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: array
                  description: "Réponses — bloquées si statut ∈ LOCKED_STATUSES"
                  items:
                    type: object
                    properties:
                      questionId:
                        type: string
                      value:
                        description: string | number | boolean | array
                status:
                  type: string
                  enum: [assigned, in_progress, submitted, reviewed, signed_evaluatee, signed_manager, signed_hr, validated, expired, archived]
                  description: |
                    Transitions autorisées par rôle :
                    employee : assigned→in_progress, in_progress→submitted, reviewed→signed_evaluatee
                    manager  : submitted→reviewed, signed_evaluatee→signed_manager
                    hr       : reviewed|signed_evaluatee|signed_manager → signed_hr
                    admin    : toutes les transitions valides
                score:
                  type: number
                  minimum: 0
                  maximum: 100
                  description: Manager/director/admin/hr uniquement
                reviewerComment:
                  type: string
                  maxLength: 5000
                  description: Manager/director/admin/hr uniquement
                nextObjectives:
                  type: string
                  maxLength: 5000
                  description: Manager/director/admin/hr uniquement
                objectiveRatings:
                  type: object
                  description: "Manager/director/admin/hr — { [questionId]: rating }"
                evaluateeComment:
                  type: string
                  maxLength: 5000
                  description: Évalué ou admin/hr uniquement
      responses:
        '200':
          description: Évaluation modifiée retournée peuplée complète
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evaluation'
        '400':
          description: Validation échouée ou transition de statut invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Accès refusé ou champ non autorisé pour ce rôle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Évaluation introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Réponses verrouillées (évaluation déjà soumise)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...